Gate the build of example requiring defmt with a feature.
Examples using defmt require the inclusion of defmt.x linker script.
However adding this file for all examples prevents those not using defmt
from being loading using probe_run.

Ideally we would want to only add defmt.x for the examples requiring it but
it is not possible for the time being.